requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-WTF==1.2.1
Flask-Bootstrap==3.3.7.1
Flask-Session==0.5.0
python-dotenv==1.0.0
PyJWT==2.10.1
email-validator==2.1.0.post1
psycopg2-binary==2.9.6
supabase==1.2.0
python-jose==3.3.0
requests==2.31.0
openai==1.6.1
Werkzeug==2.3.7
gunicorn==21.2.0  # For production deployment
uuid==1.30  # For UUID support 